# internal/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Docveta API
  version: '1'
  description: 'HTTP API of Docveta, a self-hosted document manager.


    Errors are RFC 9457 problem details (`application/problem+json`). Cookie-authenticated unsafe requests must be same-origin
    (CSRF protection); bearer tokens are exempt.


    This file is the contract: `go test ./internal/api` fails when a route is added or removed without updating it, and the
    integration test validates responses against it.'
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
- url: /
security:
- sessionCookie: []
- bearerToken: []
...
paths:
...
  /api/v1/documents:
    get:
      summary: List and search documents
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResult'
        default: *id001
      parameters:
      - name: q
        in: query
        schema: *id002
        description: Search query (supports `tag:`, `from:`, `type:`, `date:`, quotes and `-negation`)
      - name: sort
        in: query
        schema: *id002
      - name: cursor
        in: query
        schema: *id002
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - name: total
        in: query
        schema: *id007
        description: Set to false to skip counting
      - name: space_id
        in: query
        schema: *id002
        description: Comma-separated ids; repeatable
      - name: tag_id
        in: query
        schema: *id002
        description: All of these tags
      - name: any_tag_id
        in: query
        schema: *id002
      - name: not_tag_id
        in: query
        schema: *id002
      - name: correspondent_id
        in: query
        schema: *id002
      - name: document_type_id
        in: query
        schema: *id002
      - name: date_from
        in: query
        schema:
          type: string
          format: date
      - name: date_to
        in: query
        schema:
          type: string
          format: date
      - name: added_from
        in: query
        schema: *id002
        description: RFC 3339 or YYYY-MM-DD
      - name: added_to
        in: query
        schema: *id002
      - name: inbox
        in: query
        schema: *id007
      - name: status
        in: query
        schema: *id002
        description: Comma-separated statuses
      - name: untagged
        in: query
        schema: *id007
      - name: no_correspondent
        in: query
        schema: *id007
      - name: no_document_type
        in: query
        schema: *id007
      - name: trash
        in: query
        schema: *id007
      - name: mode
        in: query
        schema:
          type: string
          enum:
          - keyword
          - semantic
          - hybrid
        description: Meaning-based search needs an AI provider with an embedding model
      - name: facets
        in: query
        schema:
          type: boolean
        description: Include per-tag/correspondent/type/status counts for the current filters
    post:
      summary: Upload a document
      tags:
      - documents
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default: *id001
      description: Metadata fields must come before the `file` part; the file is streamed to storage. 409 when the file is
        a duplicate (unless `allow_duplicate`).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                space_id:
                  type: string
                  format: uuid
                  description: Defaults to the personal space
                title: *id002
                tag_ids:
                  type: string
                  description: Comma-separated
                correspondent_id:
                  type: string
                  format: uuid
                document_type_id:
                  type: string
                  format: uuid
                document_date:
                  type: string
                  format: date
                language: *id002
                allow_duplicate: *id007
                source:
                  type: string
                  enum:
                  - share
                  - scan
                file:
                  type: string
                  format: binary
                  description: Must be the last part
  /api/v1/documents/bulk:
    post:
      summary: Apply an action to many documents
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResult'
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkInput'
  /api/v1/documents/search:
    post:
      summary: Search with a JSON query
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResult'
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchQuery'
  /api/v1/documents/stats:
    get:
      summary: Counts for the home page
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        default: *id001
  /api/v1/documents/suggest:
    get:
      summary: Search-as-you-type suggestions
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Suggestion'
        default: *id001
      parameters:
      - name: q
        in: query
        schema: *id002
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 20
          default: 8
  /api/v1/documents/{id}:
    get:
      summary: Get a document
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          headers: &id008
            ETag:
              schema: *id002
              description: Document version, for If-Match
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Update document metadata
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          headers: *id008
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInput'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: If-Match
        in: header
        schema: *id002
        description: Version from ETag; 412 if it changed
    delete:
      summary: Move to trash, or delete permanently
      tags:
      - documents
      responses:
        '204': *id003
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: permanent
        in: query
        schema: *id007
  /api/v1/documents/{id}/asn:
    post:
      summary: Assign the next archive serial number
      tags:
      - documents
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /api/v1/documents/{id}/file:
    get:
      summary: Download a document file (supports Range)
      tags:
      - documents
      responses:
        '200':
          description: File content
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            image/*:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: Partial content
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: kind
        in: query
        schema:
          type: string
          enum:
          - original
          - archive
          - derived
          - thumbnail
          - best
          default: original
      - name: download
        in: query
        schema:
          type: string
          enum:
          - '1'
      - name: version
        in: query
        schema:
          type: integer
          minimum: 1
        description: Original file of an earlier version
  /api/v1/documents/{id}/history:
    get:
      summary: Change history
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/HistoryEntry'
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /api/v1/documents/{id}/notes:
    get:
      summary: List notes
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Note'
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Add a note (`@[Name](user-id)` mentions notify people)
      tags:
      - documents
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - body
              properties:
                body:
                  type: string
                  maxLength: 5000
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /api/v1/documents/{id}/notes/{note}:
    delete:
      summary: Delete a note
      tags:
      - documents
      responses:
        '204': *id003
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: note
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /api/v1/documents/{id}/pages:
    get:
      summary: Pages with recognised text
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Page'
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /api/v1/documents/{id}/reprocess:
    post:
      summary: Run processing again
      tags:
      - documents
      responses:
        '204': *id003
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: force_ocr
        in: query
        schema: *id007
  /api/v1/documents/{id}/restore:
    post:
      summary: Restore from trash
      tags:
      - documents
      responses:
        '204': *id003
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /api/v1/documents/{id}/thumbnail:
    get:
      summary: Thumbnail image
      tags:
      - documents
      responses:
        '200':
          description: Thumbnail
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        default: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
  /api/v1/status:
    get:
      summary: Instance status (public)
      tags:
      - auth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default: *id001
      security: *id004
...
  /api/v1/trash:
    delete:
      summary: Permanently delete everything in the Trash you may delete
      tags:
      - documents
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyTrashResult'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-docveta_session
      description: Set by login/setup. Named `docveta_session` when the base URL is plain http.
    bearerToken:
      type: http
      scheme: bearer
      description: Personal access token (`dvt_pat_…`) created in Settings → API tokens.
    workerToken:
      type: http
      scheme: bearer
      description: Worker token (`dvt_wrk_…`) from Administration → Processing.
...
  schemas:
...
    BulkInput:
      properties:
        action:
          type: string
        ids:
          items:
            format: uuid
            type: string
          type: array
        update:
          properties:
            add_tag_ids:
              items:
                format: uuid
                type: string
              type: array
            asn:
              format: int64
              type: integer
              nullable: true
            correspondent_id:
              format: uuid
              type: string
              nullable: true
            document_date:
              type: string
              nullable: true
            document_type_id:
              format: uuid
              type: string
              nullable: true
            inbox:
              type: boolean
              nullable: true
            language:
              type: string
              nullable: true
            physical_location:
              type: string
              nullable: true
            remove_tag_ids:
              items:
                format: uuid
                type: string
              type: array
            space_id:
              format: uuid
              type: string
              nullable: true
            tag_ids:
              items:
                format: uuid
                type: string
              type: array
              nullable: true
            title:
              type: string
              nullable: true
          type: object
      type: object
    BulkResult:
      properties:
        failed:
          items:
            properties:
              id:
                format: uuid
                type: string
              message:
                type: string
            required:
            - id
            - message
            type: object
          type: array
        succeeded:
          type: integer
      required:
      - succeeded
      - failed
      type: object
...
    Document:
      properties:
        added_at:
          format: date-time
          type: string
        asn:
          format: int64
          nullable: true
          type: integer
        correspondent:
          allOf:
          - $ref: '#/components/schemas/Ref'
          nullable: true
        deleted_at:
          format: date-time
          nullable: true
          type: string
        document_date:
          nullable: true
          type: string
        document_type:
          allOf:
          - $ref: '#/components/schemas/Ref'
          nullable: true
        has_archive:
          type: boolean
        has_thumbnail:
          type: boolean
        id:
          format: uuid
          type: string
        inbox:
          type: boolean
        language:
          type: string
        matched_page:
          nullable: true
          type: integer
        mime_type:
          type: string
        note_count:
          type: integer
        original_filename:
          type: string
        owner:
          allOf:
          - $ref: '#/components/schemas/Ref'
          nullable: true
        page_count:
          nullable: true
          type: integer
        physical_location:
          type: string
        processing_error:
          type: string
        processing_stage:
          type: string
        size_bytes:
          format: int64
          type: integer
        snippet:
          items:
            properties:
              hit:
                type: boolean
              text:
                type: string
            required:
            - text
            type: object
          type: array
        source:
          type: string
        space:
          $ref: '#/components/schemas/Ref'
        status:
          type: string
        tags:
          items:
            $ref: '#/components/schemas/Ref'
          type: array
        title:
          type: string
        updated_at:
          format: date-time
          type: string
        version:
          type: integer
        has_derived:
          type: boolean
        custom_fields:
          type: array
          items:
            $ref: '#/components/schemas/CustomValue'
        suggestion_count:
          type: integer
      required:
      - id
      - space
      - title
      - document_date
      - added_at
      - updated_at
      - correspondent
      - document_type
      - tags
      - language
      - page_count
      - asn
      - physical_location
      - inbox
      - status
      - processing_stage
      - mime_type
      - size_bytes
      - original_filename
      - source
      - owner
      - has_archive
      - has_thumbnail
      - note_count
      - version
      - has_derived
      - custom_fields
      - suggestion_count
      type: object
...
    HistoryEntry:
      properties:
        action:
          type: string
        actor:
          allOf:
          - $ref: '#/components/schemas/Ref'
          nullable: true
        actor_type:
          type: string
        created_at:
          format: date-time
          type: string
        details:
          additionalProperties:
            nullable: true
          type: object
        id:
          format: uuid
          type: string
      required:
      - id
      - actor
      - actor_type
      - action
      - details
      - created_at
      type: object
...
    ListResult:
      properties:
        items:
          items:
            allOf:
            - $ref: '#/components/schemas/Document'
            nullable: true
          type: array
        next_cursor:
          nullable: true
          type: string
        total:
          nullable: true
          type: integer
        mode:
          type: string
          enum:
          - keyword
          - semantic
          - hybrid
        facets:
          $ref: '#/components/schemas/Facets'
      required:
      - items
      - next_cursor
      type: object
...
    Stats:
      properties:
        added_this_week:
          type: integer
        bytes:
          format: int64
          type: integer
        failed:
          type: integer
        inbox:
          type: integer
        processing:
          type: integer
        total:
          type: integer
        trash:
          type: integer
        ocr_available:
          type: boolean
        trash_retention_days:
          type: integer
      required:
      - total
      - inbox
      - processing
      - failed
      - trash
      - added_this_week
      - bytes
      - ocr_available
      - trash_retention_days
      type: object
...
    Suggestion:
      type: object
      required:
      - id
      - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
...